pyile_manager_backend/pyproject.toml
[project]
name = "pyile-manager"
version = "0.1.0"
description = "Automated File Manger written in python!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "watchdog>=6.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pdfplumber>=0.11.0",
    "python-pptx>=1.0.0",
    "pillow>=11.0.0",
    "pyinstaller>=6.11.0",
]

[tool.ruff]
line-length = 120 # for lint
indent-width = 4
# target-version = "py311"

[tool.ruff.lint.pycodestyle]
max-line-length = 120 # for format

[tool.ruff.lint]
# line-length = 120 # don't use in tool.ruff.lint
# - E: pycodestyle errors (PEP 8 style errors)
# - F: Pyflakes (logical errors, unused imports, etc.)
# - I: isort (import sorting)
# - UP: pyupgrade (suggests syntax upgrades for newer Python versions)
# - B: flake8-bugbear (likely bugs and design problems)
select = ["E", "F", "UP", "B"]
ignore = [
  # "E501", # Ignore line length errors if desired
  "F401", # Ignore unused imports
  "UP015",
]

[tool.ruff.format]
quote-style = "double"
# quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
